+++ process/block/pendingMb/pendingMiniBlocks_test.go
@@ -1,4 +1,4 @@
-package metachain
+package pendingMb
 
 import (
 	"testing"
